Update valuations.tex
On the comment between lemmas 13.2.8 and 9: the definition of order suggest it should be n rather than p^n. 

On the proof of Lemma 13.2.10, second and third lines of the mathematical expression |1+a|^n: we can add |a|^j (2nd line) and n+1 rather than n (3rd line).

On the proof of Ostrowski theorem 13.3.2's proof - Archimedean case: On the first line of the inequality of |b|, there is an absolute value missing. Later, right after the inequality of |c| involving only a and c, it was written |\cdot| is nonarchimedean when we are in the case archimedean.

On the definition of |\cdot|_\infty for k(t), you mention it differs from the one in Cassels in two ways: you mention the first one |c| > 1 but fail to mention the exchange of deg(u) - deg(v) rather than its negative, which makes it totally equivalent. (I haven't made any changes here.)

In your explanation of the choice of c > 1 you wrote |a+1| = ... c^1 < 1 when it should be c^{-1} > 1. And later you wrote |p| = c^1 > 1 when it should be |p| = c^{-1} < 1.